+++ src/MiloneTranslation/MonoTyIr.fs
@@ -42,11 +42,13 @@ type MonoTy =
   // Nominal types.
   | UnionMt of unionTy: TySerial
   | RecordMt of recordTy: TySerial
+  | OpaqueMt of opaqueTy: TySerial
 
 [<NoEquality; NoComparison>]
 type TyDef =
   | UnionTyDef of Ident * tyArgs: TySerial list * VariantSerial list * Loc
   | RecordTyDef of Ident * fields: (Ident * Ty * Loc) list * IsCRepr * Loc
+  | OpaqueTyDef of Ident * Loc
 
 [<RequireQualifiedAccess; NoEquality; NoComparison>]
 type VarDef =