+++ src/MiloneTranslationTypes/HirTypes.fs
@@ -59,6 +59,7 @@ type Tk =
   | MetaTk of metaTy: TySerial * metaLoc: Loc
   | UnionTk of unionTy: TySerial
   | RecordTk of recordTy: TySerial
+  | OpaqueTk of opaqueTy: TySerial
 
 /// Type of expressions.
 [<Struct; NoEquality; NoComparison>]
@@ -76,6 +77,8 @@ type TyDef =
   | UnionTyDef of Ident * tyArgs: TySerial list * VariantSerial list * Loc
 
   | RecordTyDef of Ident * fields: (Ident * Ty * Loc) list * IsCRepr * Loc
+
+  | OpaqueTyDef of Ident * Loc
 
 /// Definition of named value in HIR.
 [<RequireQualifiedAccess; NoEquality; NoComparison>]