+++ tests/primitives/native_type/native_type.fs
@@ -1,14 +1,15 @@
 module rec native_type.Program
 
-// __nativeType<``ANY C TYPE``> is expanded to `ANY C TYPE` in C.
+// - __nativeType<``ANY C TYPE``> is expanded to `ANY C TYPE` in C.
+// - Opaque types are compiled to incomplete struct types in C.
 // See also x_native_code.md in docs.
 
 open Std.Ptr
 
 type FILE = __nativeType<FILE>
 
-// It can include spaces.
-type OpaqueStruct = __nativeType<``struct Opaque``>
+[<Opaque>]
+type OpaqueStruct = private | OS
 
 let strAsPtr (s: string) : InPtr<char> = __nativeFun ("string_as_ptr", s)